+++ semcore/data-table/src/components/Body/Cell.types.ts
@@ -1,3 +1,4 @@
+import type { Intergalactic } from '@semcore/ui/core';
 import type * as React from 'react';
 
 import type { DTRow, DTRows } from './Row.types';
@@ -6,10 +7,10 @@ import type { DTColumn } from '../Head/Column.types';
 
 export type Theme = 'muted' | 'info' | 'success' | 'warning' | 'danger';
 
-export type DataTableCellProps = {
+export type DataTableCellProps<DataKeyType> = {
   id: string;
   accordionId: string;
-  row: DTRow;
+  row: DTRow<DataKeyType>;
   rowIndex: number;
   column: DTColumn;
   columnIndex: number;
@@ -22,14 +23,19 @@ export type DataTableCellProps = {
   isAccordionRow?: boolean;
   animationExpand?: boolean;
   accordionRowIndex?: number;
-  rows: DTRows;
+  rows: DTRows<DataKeyType>;
 };
 
-export type CellPropsInner = {
+export type CellPropsInner<DataKeyType> = {
   use: DTUse;
 
   virtualScroll: boolean;
   tableRef: React.RefObject<HTMLDivElement>;
   accordionDuration?: number | [number, number];
-  onClick: (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow }) => void;
+  onClick: (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow<DataKeyType> }) => void;
 };
+
+export type DataTableCellType = (<DataKeyType, Tag extends Intergalactic.Tag = 'div'>(
+  props: Intergalactic.InternalTypings.ComponentProps<Tag, 'div', DataTableCellProps<DataKeyType>>
+) => Intergalactic.InternalTypings.ComponentRenderingResults) &
+Intergalactic.InternalTypings.ComponentAdditive<'div', 'div', DataTableCellProps<any>>;