+++ semcore/data-table/src/components/Body/Row.types.ts
@@ -1,31 +1,31 @@
+import type { Intergalactic } from '@semcore/ui/core';
+
 import type { DataTableCellProps } from './Cell.types';
 import type { MergedColumnsCell, MergedRowsCell } from './MergedCells';
 import type { ACCORDION, ROW_GROUP, ROW_INDEX, UNIQ_ROW_KEY } from '../DataTable/DataTable';
 import type { DTValue, DTUse, DataTableData } from '../DataTable/DataTable.types';
 import type { DTColumn } from '../Head/Column.types';
 
-export type UniqRowKey = string;
-
-export type DTRow = {
-  [UNIQ_ROW_KEY]: UniqRowKey;
+export type DTRow<DataKeyType> = {
+  [UNIQ_ROW_KEY]: DataKeyType;
   [ROW_INDEX]: number;
   [key: string]: DTValue | MergedRowsCell | MergedColumnsCell;
   [ACCORDION]?: React.ReactNode | DataTableData | undefined;
-  [ROW_GROUP]?: Set<UniqRowKey>;
+  [ROW_GROUP]?: Set<DataKeyType>;
 };
-export type DTRows = Array<DTRow | DTRow[]>;
+export type DTRows<DataKeyType> = Array<DTRow<DataKeyType> | DTRow<DataKeyType>[]>;
 
-export type DataTableRowProps = {
-  row: DTRow;
+export type DataTableRowProps<DataKeyType> = {
+  row: DTRow<DataKeyType>;
   mergedRow?: boolean;
 
-  isAccordionRow?: DataTableCellProps['isAccordionRow'];
-  animationExpand?: DataTableCellProps['animationExpand'];
-  accordionRowIndex?: DataTableCellProps['accordionRowIndex'];
+  isAccordionRow?: DataTableCellProps<DataKeyType>['isAccordionRow'];
+  animationExpand?: DataTableCellProps<DataKeyType>['animationExpand'];
+  accordionRowIndex?: DataTableCellProps<DataKeyType>['accordionRowIndex'];
   isNonInteractive?: boolean;
 };
 
-export type RowPropsInner = JSX.IntrinsicElements['div'] & {
+export type RowPropsInner<DataKeyType> = JSX.IntrinsicElements['div'] & {
   use: DTUse;
   /**
    * Expanded flag for rows with accordion
@@ -39,24 +39,24 @@ export type RowPropsInner = JSX.IntrinsicElements['div'] & {
   mergedRow?: boolean;
 
   columns: DTColumn[];
-  row: DTRow | DTRow[];
-  rows: DTRows;
+  row: DTRow<DataKeyType> | DTRow<DataKeyType>[];
+  rows: DTRows<DataKeyType>;
   rowIndex: number; // from 0
   ariaRowIndex: number; // from 1 + 1 header
   gridRowIndex: number; // from 1 + 1 (or 2 if it has group) header
 
-  expandedRows: Set<string>;
-  onExpandRow: (expandedRow: DTRow) => void;
+  expandedRows: Set<DataKeyType>;
+  onExpandRow: (expandedRow: DTRow<DataKeyType>) => void;
 
   gridTemplateAreas: string[];
   gridTemplateColumns: string[];
   accordionDataGridArea: string;
 
-  selectedRows?: UniqRowKey[];
+  selectedRows?: DataKeyType[];
   onSelectRow?: (
     isSelect: boolean,
     selectedRowIndex: number,
-    row: DTRow,
+    row: DTRow<DataKeyType>,
     event?: React.SyntheticEvent<HTMLElement>,
   ) => void;
 
@@ -72,3 +72,8 @@ export type RowPropsInner = JSX.IntrinsicElements['div'] & {
     cell: Pick<DTColumn, 'name' | 'fixed'>,
   ) => [side: 'left' | 'right', style: string | number] | [side: undefined, style: undefined];
 };
+
+export type DataTableRowType = (<DataKeyType, Tag extends Intergalactic.Tag = 'div'>(
+  props: Intergalactic.InternalTypings.ComponentProps<Tag, 'div', DataTableRowProps<DataKeyType>>
+) => Intergalactic.InternalTypings.ComponentRenderingResults) &
+Intergalactic.InternalTypings.ComponentAdditive<'div', 'div', DataTableRowProps<any>>;