+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -27,7 +27,7 @@ import { localizedMessages } from '../../translations/__intergalactic-dynamic-lo
 import { Body } from '../Body/Body';
 import type { BodyPropsInner } from '../Body/Body.types';
 import { MergedColumnsCell, MergedRowsCell } from '../Body/MergedCells';
-import type { DTRow, UniqRowKey } from '../Body/Row.types';
+import type { DTRow } from '../Body/Row.types';
 import type { DataTableColumnProps, DTColumn } from '../Head/Column.types';
 import type { DataTableGroupProps } from '../Head/Group.type';
 import { Head } from '../Head/Head';
@@ -47,13 +47,17 @@ type State = {
   selectAllMessage: string;
 };
 
-class DataTableRoot<D extends DataTableData> extends Component<
-  DataTableProps<D>,
-  {},
-  {},
+class DataTableRoot<
+  Data extends DataTableData,
+  DataKey extends keyof Data[number],
+  DataKeyType extends Data[number][DataKey],
+> extends Component<
+    DataTableProps<Data, DataKey, DataKeyType>,
+    {},
+    {},
   typeof DataTableRoot.enhance,
-  { use: DTRow; expandedRows: Set<string>; renderEmptyData: () => React.ReactNode }
-> {
+  { use: DTRow<DataKeyType>; expandedRows: Set<DataKeyType>; renderEmptyData: () => React.ReactNode }
+  > {
   static displayName = 'DataTable';
   static style = style;
 
@@ -84,14 +88,14 @@ class DataTableRoot<D extends DataTableData> extends Component<
   private gridAreaGroupMap = new Map<number, string>();
 
   private columnsSplitter = '/';
-  private rows: Array<DTRow | DTRow[]> = [];
-  private flatRows: DTRow[] = [];
+  private rows: Array<DTRow<DataKeyType> | DTRow<DataKeyType>[]> = [];
+  private flatRows: DTRow<DataKeyType>[] = [];
 
   private selectAllMessageTimer = 0;
 
   private headerNodesMap = new Map();
 
-  constructor(props: DataTableProps<D>) {
+  constructor(props: DataTableProps<Data, DataKey, DataKeyType>) {
     super(props);
 
     if (props.children) {
@@ -197,7 +201,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     };
   }
 
-  getHeadProps(): HeadPropsInner<D> {
+  getHeadProps(): HeadPropsInner<Data, DataKey, DataKeyType> {
     const {
       use,
       compact,
@@ -243,7 +247,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     };
   }
 
-  getBodyProps(): BodyPropsInner {
+  getBodyProps(): BodyPropsInner<DataKeyType> {
     const {
       use,
       compact,
@@ -297,7 +301,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     };
   }
 
-  handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow }) => {
+  handleCellClick = (e: React.SyntheticEvent, opt: { rowIndex: number; colIndex: number; row?: DTRow<DataKeyType> }) => {
     if (lastInteraction.isMouse()) {
       console.log('click on cell in data table handler', opt);
       this.initFocusableCell([this.hasFocusableInHeader() ? opt.rowIndex + 1 : opt.rowIndex, opt.colIndex]);
@@ -307,7 +311,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
   handleSelectRow = (
     isSelected: boolean,
     selectedRowIndex: number,
-    row: DTRow,
+    row: DTRow<DataKeyType>,
     event?: React.SyntheticEvent<HTMLElement>,
   ) => {
     const { selectedRows, onSelectedRowsChange } = this.asProps;
@@ -372,7 +376,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
       this.columns.some((column) => column.sortable);
   };
 
-  onExpandRow = (expandedRow: DTRow) => {
+  onExpandRow = (expandedRow: DTRow<DataKeyType>) => {
     const { expandedRows } = this.asProps;
     if (expandedRows.has(expandedRow[UNIQ_ROW_KEY])) {
       expandedRows.delete(expandedRow[UNIQ_ROW_KEY]);
@@ -1070,12 +1074,12 @@ class DataTableRoot<D extends DataTableData> extends Component<
     return [calculatedColumns, treeColumns];
   }
 
-  private calculateRows(): Array<DTRow[] | DTRow> {
+  private calculateRows(): Array<DTRow<DataKeyType>[] | DTRow<DataKeyType>> {
     const columns = this.columns;
     // @ts-ignore
     const { data, uid, uniqueRowKey } = this.props;
 
-    const rows: Array<DTRow[] | DTRow> = [];
+    const rows: Array<DTRow<DataKeyType>[] | DTRow<DataKeyType>> = [];
     const columnNames = columns.map((column: DTColumn) => column.name);
 
     let rowIndex = 0;
@@ -1084,7 +1088,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
 
     const makeDtRow = (row: DataRowItem, excludeColumns?: string[]) => {
       const columns = new Set(columnNames);
-      const dtRow = Object.entries(row).reduce<DTRow>(
+      const dtRow = Object.entries(row).reduce<DTRow<DataKeyType>>(
         (acc, [key, value]) => {
           const columnsToRow = key.split(this.columnsSplitter);
 
@@ -1108,8 +1112,12 @@ class DataTableRoot<D extends DataTableData> extends Component<
           return acc;
         },
         {
-          [UNIQ_ROW_KEY]:
-            row[UNIQ_ROW_KEY] || (uniqueRowKey ? row[uniqueRowKey] : `${uid}_${(rowIndex + id).toString(36)}`) as UniqRowKey,
+          /*
+            row -> DataRowItem
+            uniqueRowKey is a `keyof Data[number]` -> `keyof DataRowItem`
+          */
+          // @ts-ignore
+          [UNIQ_ROW_KEY]: row[UNIQ_ROW_KEY] || (uniqueRowKey ? row[uniqueRowKey] : `${uid}_${(rowIndex + id).toString(36)}`),
           [ROW_INDEX]: rowIndex,
         },
       );
@@ -1131,10 +1139,10 @@ class DataTableRoot<D extends DataTableData> extends Component<
       const groupedRows: DataTableData | undefined = row[ROW_GROUP];
 
       if (groupedRows) {
-        const innerRows: DTRow[] & { [ACCORDION]?: React.ReactElement } = [];
+        const innerRows: DTRow<DataKeyType>[] & { [ACCORDION]?: React.ReactElement } = [];
 
         const groupedKeys: string[] = [];
-        const groupedRowData = Object.entries(row).reduce<Omit<DTRow, symbol>>(
+        const groupedRowData = Object.entries(row).reduce<Omit<DTRow<DataKeyType>, symbol>>(
           (acc, [key, value]) => {
             acc[key] = new MergedRowsCell(value, groupedRows.length, row[ACCORDION]);
             groupedKeys.push(key);
@@ -1147,7 +1155,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
         );
 
         groupedRows.forEach((childRow, index) => {
-          let dtRow: DTRow;
+          let dtRow: DTRow<DataKeyType>;
           if (index === 0) {
             const rowData = {
               ...childRow,